Displays the results of an \(E_n\)-test for two values with extended uncertainties

The function displays the results of a \(E_n\)-test for two values with extended uncertainty. The returned text is suitable for the SI confronta shiny app.

fct_entest_2values_unc(data, response, uncertainty, group)

Arguments

data

a data.frame or data.table with the results relevant for testing. A single-level grouping factor variable and two numeric vectors with the measurements values and their uncertainties should be included.

response

the name of a numeric vector in data. Quotation (" ") is not required.

uncertainty

the name of a numeric vector in data. Quotation (" ") is not required.

group

the name of a single-level factor variable that identifies the group in data. Quotation (" ") is not required.

Value

A list with the following items:

hypotheses

a named vector of strings, being h0 and h1 the null and alternative hypothesis, respectively.

difference

a named vector of numbers, being mean, lwrci and uprci the mean, lower and upper ends of the confidence interval for the difference of the two values, respectively. The confidence interval is calculated considered the combined extended uncertainties of the two values.

test

a named vector of numbers, being entest and ecritical the value of the test statistic and the associated critical value, respectively.

result

a string indicating whether H0 should be rejected or not.

Details

\(E_n\)-test is calculated with the following equation: $$E_n = \frac{|x - y|}{\sqrt{U(x)^2 + U(y)^2}}$$ for \(E_n \leq 1\), the difference of the two values \(x\) and \(y\) is accounted by their extended uncertainties (\(U\)) and the null hypothesis is not rejected. When \(E_n > 1\) the two values are different and the null hypothesis is rejected.