Displays the results of an \(F\)-test for two groups of values

The function displays the results of an \(F\)-test performed on two groups of values. The returned text is suitable for the SI confronta shiny app.

fct_ftest_2samples(
  data,
  response,
  group,
  significance = 0.95,
  alternative = "different"
)

Arguments

data

a data.frame or data.table with the results relevant for testing. At least a two-levels grouping factor variable and a numeric vector with the measurements should be included.

response

the name of a numeric vector in data. Quotation (" ") is not required.

group

the name of a two-level factor variable that identifies the groups in data. Quotation (" ") is not required.

significance

a number, typically either 0.90, 0.95 (default) or 0.99 indicating the confidence level for the test.

alternative

a character string specifying the alternative hypothesis, must be one of "different" or "greater".

Value

A list with the following items:

hypotheses

a named vector of strings, being h0 and h1 the null and alternative hypothesis, respectively.

ratio

a named vector of numbers, being mean, lwrci and uprci the ratio of the variances of the two groups and the lower and upper ends of the confidence interval, respectively. The confidence interval is calculated considering both the significance and alternative arguments. For alternative = "greater" only the lower end of the confidence interval will be calculated.

test

a named vector of numbers, being dof, fsper, ftheo and pvalue the degrees of freedom, the calculated value of the \(F\) statistic, the tabulated value of the \(F\) statistic and the \(p\)-value associated to the test.

result

a string indicating whether H0 should be rejected or not.

Details

\(F\)-test is calculated using the base-R function var.test, with the group with the numerically larger variance as first argument. As a consequence, for the variances of two series of values \(A\) and \(B\), being the first one numerically greater than the second one, the alternative hypothesis tested can only be \(\mathrm{Var}(A) \neq \mathrm{Var}(B)\) (alternative = "different") or \(\mathrm{Var}(A) > \mathrm{Var}(B)\) (alternative = "greater").